requirements.txt
absl-py==2.3.0
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.45.3
blinker==1.9.0
blobfile==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
community==1.0.0b1
contourpy==1.3.1
cpgclientlib==0.11.321
cycler==0.12.1
datasets==3.3.2
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
exceptiongroup==1.2.2
fairscale==0.4.13
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.20.3
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
kiwisolver==1.4.8
llama_models==0.0.63
llama_stack==0.0.63
llama_stack_client==0.0.63
lxml==5.3.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.57.0
packaging==24.2
pandas==2.2.3
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==6.30.2
psutil==6.1.0
ptyprocess==0.7.0
pyaml==24.12.1
pyarrow==19.0.1
pycryptodomex==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.4
Pygments==2.18.0
pygraphviz==1.14
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-louvain==0.16
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.29.1
setproctitle==1.3.6
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.3
torch==2.5.0+cu124
torch-geometric==2.6.1
torchaudio==2.5.0+cu124
torchvision==0.20.0+cu124
tqdm==4.67.0
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.2
xxhash==3.5.0
yarl==1.18.3
